refactor(userspace-dp): extract helpers/persistence module
Continue the #6234 decomposition: move the state-file persistence domain
out of server/helpers.rs into helpers/persistence.rs (third of four
submodules).

This carries the #5469 lock discipline intact: write_state holds the
ServerState lock only long enough to refresh_status and clone the minimal
owned OwnedStatePayload (build_state_payload, the sole guard-touching
half), then drops the guard before the fallible encode() + StateWriter
persist. The OwnedStatePayload type structure — owning its data, unable
to reach the guard — remains the compile-time proof that serialization
is lock-free. The #5469 pre-persist lock probe
(record/clear/take_pre_persist_lock_free) moves with it, so the
write_state_releases_lock_before_persist fail-on-revert guard still
binds.

Pure code-motion: bodies byte-for-byte identical. Explicit imports
(crate::protocol status/snapshot types, crate::server::ServerState,
serde::Serialize, std::sync) replace the crate-root glob; the sole
cross-module dep routes through the facade (super::refresh_status).
mod.rs re-exports persistence::* so write_state keeps its
server::helpers::write_state call-site path in the handler dispatch tail.

Validation: `cargo build` on userspace-dp is green with no new warnings
in helpers/.

Refs #6234
